mig to ts for Link component and edit link container
